[Update] id 자료형 int로 변경

+++ src/main/java/com/first/bulletinboard/repository/UserRepository.java
@@ -1,10 +1,10 @@
 package com.first.bulletinboard.repository;
 
-import com.first.bulletinboard.domain.User;
+import com.first.bulletinboard.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUserName(String userName);
 }
